Improved the performance of GetNeighbors and GetClosedNetwork.
If we have an outlink to a peer, they might be a neighbor if they have an outlink to us. There's no point in getting the peer's outlinks, since we just need to query our own inlinks to find out if we're two-way with that peer.

Using this, we can also cut down on the number of link checks we make, in GetClosedNetwork(), because we can filter out seen peers before we bother to check that we're two-way.